use longs in blingers
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/src@20625 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ uvm-lib/impl/com/untangle/uvm/engine/MessageManagerImpl.java
@@ -462,18 +462,18 @@ class MessageManagerImpl implements LocalMessageManager
             readProcFile("/proc/meminfo", MEMINFO_PATTERN, MEMINFO_KEEPERS, m,
                          1024);
 
-            Integer memFree = (Integer)m.get("MemFree");
+            Long memFree = (Long)m.get("MemFree");
             if (null == memFree) {
-                memFree = 0;
+                memFree = 0L;
             }
 
-            Integer i = (Integer)m.get("Cached");
+            Long i = (Long)m.get("Cached");
             if (null != i) {
                 memFree += i;
             }
             m.remove("Cached");
 
-            i = (Integer)m.get("Buffers");
+            i = (Long)m.get("Buffers");
             if (null != i) {
                 memFree += i;
             }
